requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
APScheduler==3.10.4
anthropic==0.18.1
python-dotenv==1.0.0
gunicorn==21.2.0
